+++ packages/asset-swapper/CHANGELOG.md
@@ -5,7 +5,19 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
-## v6.8.0 - _October 19, 2021_
+## v16.32.0 - _November 9, 2021_
+
+    * Extended Quote Report (#361)
+
+## v16.31.0 - _November 3, 2021_
+
+    * Added `Curve`, `Curve_V2` and `KyberDmm` to Avalanche (#363)
+
+## v16.30.1 - _November 3, 2021_
+
+    * Dependencies updated
+
+## v16.30.0 - _October 19, 2021_
 
     * Fantom deployment (#347)